+++ TerminalGameEngine/PongBar.cpp
@@ -6,8 +6,8 @@ PongBar::PongBar
 (
     int xPos,
     int yPos,
-    uint sizeX,
-    uint sizeY,
+    size_t sizeX,
+    size_t sizeY,
     char modelChar,
     double moveSpeed,
     double deflectBallFactor,
